envirocare-backend-main/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
boto3==1.35.67
botocore==1.35.67
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2023.1
cssselect2==0.7.0
distro==1.9.0
Django==5.1.7
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-redis==5.4.0
django-rest-swagger==2.2.0
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
exponent_server_sdk==2.1.0
fonttools==4.53.1
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
kombu==5.3.4
Markdown==3.5.1
MarkupSafe==2.1.3
multidict==6.0.5
numpy==2.1.1
openai==1.69.0
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.16.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.8
requests==2.31.0
s3transfer==0.10.2
sendgrid==6.11.0
setuptools==75.3.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tinycss2==1.3.0
tqdm==4.67.1
twilio==9.2.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.8.2
yarl==1.9.4
zopfli==0.2.3
